refactor: auth, common/jwt 패키지 병합, JwtTokenProvider & AuthProvider -> AuthProvider
