+++ src/main/java/uk/jinhy/survey_mate_api/auth/presentation/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package uk.jinhy.survey_mate_api.common.jwt;
+package uk.jinhy.survey_mate_api.auth.presentation;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
